Including some definitions.
I'm not sure how nuanced you want these definitions to get, but YSA frame the ways youth can serve in four ways. Two of those are here, service (volunteering), and advocacy. The other two are awareness and philanthropy. I included the definitions in the body for the other contributors and decision makers to review. 

I am also fairly new to this, so if this was the wrong way to go about proposing change, let me know.